+++ src/components/Footer/Footer.css
@@ -1,8 +1,7 @@
 footer {
-  margin-top: 35vh;
   width: 100vw;
-  position:relative;
-  bottom: 0px;
+  position:fixed;
+  bottom: 0;
   padding: 20px;
   text-align: center;
   color: orange;
