+++ src/components/UserPage/UserPage.css
@@ -1,8 +1,7 @@
 .user-page {
     display: flex;
     justify-content: space-around;
-    margin-left: 20px;
-    margin-right: 20px;
+    padding: 30px;
 }
